+++ Bud.Core/src/main/cs/Cli/CommandListParser.cs
@@ -4,8 +4,6 @@ using System.Collections.Immutable;
 
 namespace Bud.Cli {
   public static class CommandListParser {
-    public const string OptionsAndCommandsSeparator = "--";
-
     public static IEnumerable<Command> ToCommandList(IEnumerable<string> commandLineArguments) {
       if (commandLineArguments == null) {
         return ImmutableList<Command>.Empty;
@@ -54,15 +52,15 @@ namespace Bud.Cli {
     }
 
     public static void ExtractOptionsAndCommands(string[] commandLineArgs, out string[] options, out string[] commands) {
-      int indexOfSeparator = Array.FindIndex(commandLineArgs, arg => OptionsAndCommandsSeparator.Equals(arg));
+      int indexOfSeparator = Array.FindIndex(commandLineArgs, MacroCommand.IsMacroCommand);
       if (indexOfSeparator < 0) {
         options = commandLineArgs;
         commands = new string[] {};
       } else {
         options = new string[indexOfSeparator];
         Array.Copy(commandLineArgs, options, options.Length);
-        commands = new string[commandLineArgs.Length - indexOfSeparator - 1];
-        Array.Copy(commandLineArgs, indexOfSeparator + 1, commands, 0, commands.Length);
+        commands = new string[commandLineArgs.Length - indexOfSeparator];
+        Array.Copy(commandLineArgs, indexOfSeparator, commands, 0, commands.Length);
       }
     }
   }