+++ RELEASE_NOTES.md
@@ -1,4 +1,4 @@
-## Upcoming release
+## 0.3.0
 
 - __New feature__: Introduced the `@watch` macro. This macro watches sources in the `src` folder and evaluates given tasks whenever the sources change. Example of usage: